perf: Mv balance history computations to HistoryChart
